Abstract the process of downloading and parsing the json files from github
